Remove MeasurementUnit and related references
Remove the MasurementUnit entity and all related UI, repo, DTO and DB references. Deleted the MeasurementUnit class, seeding and measurement unit forms; removed measurement unit properties and navigation from Product, InventoryDbContext and repositories; updated migrations (renamed InitialCommit -> NewInitialCommit) to drop the MasurementUnits table/columns and adjusted constraints; removed the checked-in inventory.db. Also updated design-time DbContext factory to store inventory.db in CommonApplicationData, adjusted DTOs and UI controls (product/invoice views and product forms) to no longer show or expect unit fields, and renamed/cleaned product add/update form types accordingly.

+++ LogicLayer/Services/Invoices/InvoiceService.cs
@@ -93,7 +93,7 @@ namespace LogicLayer.Services.Invoices
 
             return new InvoiceReadDto()
             {
-                InvoiceId = Invoice.InvoiceId,
+                InvoiceId = (int)Invoice.InvoiceId,
                 OpenDate = Invoice.OpenDate,
                 CloseDate = Invoice.CloseDate,
                 InvoiceType = Invoice.InvoiceType.GetDisplayName(),
@@ -127,7 +127,7 @@ namespace LogicLayer.Services.Invoices
 
             return new InvoiceListDto()
             {
-                InvoiceId = invoice.InvoiceId,
+                InvoiceId = (int)invoice.InvoiceId,
                 OpenDate = invoice.OpenDate,
                 CloseDate = invoice.CloseDate,
                 InvoiceType = invoice.InvoiceType.GetDisplayName(),
